requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
mysql-connector-python==9.2.0
cryptography==42.0.5
PyMySQL==1.1.0
python-dotenv==1.0.1
